Refactor menu, update readme, new unit test for numbers in a word
