+++ WordParser/Program.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 
 namespace WordParser
 {
@@ -9,28 +8,9 @@ namespace WordParser
         {
             var sentenceParser = new SentenceParser();
 
-            if (args.Contains("-i"))
+            if (args.Length == 1 && args[0] == "-i")
             {
-                bool repeatInput = true;
-
-                while (repeatInput)
-                {
-                    var menuOption = true;
-                    Console.WriteLine("Enter a sentence to be parsed: ");
-                    var input = Console.ReadLine();
-
-                    Console.WriteLine(sentenceParser.ParseWords(input));
-
-                    while (menuOption)
-                    {
-                        Console.WriteLine("\nSelect an option: \"q\" - quit; \"c\" - continue parsing more sentences");
-                        input = Console.ReadLine().TrimEnd();
-                        repeatInput = input == "c";
-                        menuOption = input != "c" && input != "q";
-                    }
-                }
-
-                Console.WriteLine("Goodbye.");
+                new Menu().OpenInteractiveMenu();
             }
             else
             {
